fix: 멀티샷 발사 수에 따라 한 발당 피해가 줄어들도록 수정
- 발당 피해 1발 100%, 2발 70%, 3발 55%, 4발 45%, 5발 40% (총 피해 최대 2배)
- 소수 피해가 반올림되지 않도록 총알 피해와 적 체력을 float로 변경
- 레벨업 카드, 능력 툴팁, 상점 공격력 항목에 한 발당 피해 표시

+++ Assets/Scripts/EnermyController.cs
@@ -5,7 +5,7 @@ public class EnermyController : MonoBehaviour
 {
     public float speed = 2f;
     public int setEnemyHP = 10;
-    public int EnemyHealth;
+    public float EnemyHealth;
 
 
     private Transform player;
@@ -126,7 +126,7 @@ public class EnermyController : MonoBehaviour
         return Vector2.ClampMagnitude(push, 1.5f);
     }
 
-    public void TakeDamage(int damage, float knockBack, Vector3 dir)
+    public void TakeDamage(float damage, float knockBack, Vector3 dir)
     {
         if (isDead) return;
